redmineID #4143 Ironmq is now izy'ized.
You will need to create a build.properties for both amazon-sqs and ironmq that contains 2 properties;

ivy.user=deployment

ivy.password=<ask for password>

HornetQ does not yet have a build file - so we'll finish this one after that is done.

SwiftMQ is under construction - we'll ivy'ize this one once it's complete.